--- server/sensor/sensor_server_test.go
@@ -1,1 +1,0 @@
-package sensor
