--- ui/src/app/sensors/components/sensor-list/sensor-list.scss
@@ -1,8 +1,0 @@
-@import 'node_modules/argo-ui/src/styles/config';
-
-.argo-table-list {
-    margin: 1em;
-    &__row:hover {
-        box-shadow: 1px 2px 3px rgba($argo-color-gray-9, 0.1), 0 0 0 1px rgba($argo-color-teal-5, 0.5);
-    }
-}
